refactor: 분리되어 있던 createFormItemLabel을 Input class로 이동
