--- src/components/createFormItemLabel.js
@@ -1,8 +1,0 @@
-export const createFormItemLabel = (type, title) => {
-  const label = document.createElement('label');
-  label.setAttribute('for', type);
-  label.classList.add('text-caption');
-  label.textContent = title;
-
-  return label;
-};
